window and server.
