+++ internal/window/serve_windows.go
@@ -1,0 +1,8 @@
+// +build windows
+
+package window
+
+func (s *Server) serve() {
+
+	// handle windows
+}
